refactor(core): lift shared callRoutine helpers to Connection
`callRoutineFunction` collapses the identical `select <qualified>(?, ?…)` path
across MySQL, MSSQL, PostgreSQL, and SQLite. `applyRoutineOutParams` shares the
OUT-`ScalarReference.set()` loop across MySQL, MSSQL, PostgreSQL, and Oracle.
Procedure SQL building stays per-driver since the dialects diverge sharply
(MySQL session @vars, MSSQL DECLARE/OUTPUT batch, PG refcursors, Oracle PL/SQL
binds).
